исправление ошибки работы тестера
