+++ js/two-cards-game-screen.js
@@ -1,4 +1,5 @@
-import {getElementFromTemplate, getLevelProgressBar, enterNextLevel} from './utils.js';
+import {getElementFromTemplate, getLevelProgressBar} from './utils.js';
+import enterNextLevel from './enter-next-level.js';
 import getHeader from './header.js';
 import getFooterMarkup from './footer.js';
 import {userState} from './utils.js';
